examples: fix some compiler warnings

+++ docs/examples/sepheaders.c
@@ -31,7 +31,7 @@
 
 static size_t write_data(void *ptr, size_t size, size_t nmemb, void *stream)
 {
-  int written = fwrite(ptr, size, nmemb, (FILE *)stream);
+  size_t written = fwrite(ptr, size, nmemb, (FILE *)stream);
   return written;
 }
